Add/upload picture 2 (v 0.09)

+++ frontend/controllers/TaskController.php
@@ -12,6 +12,7 @@ namespace frontend\controllers;
 use common\models\Condition;
 use common\models\Event;
 use frontend\components\GoogleDriveHelper;
+use frontend\components\GoogleIdentityHelper;
 use Yii;
 use yii\web\Controller;
 use common\models\Task;
@@ -150,9 +151,18 @@ class TaskController extends Controller {
                 $task->description = $taskForSave['description'];
             }
 
-            if (isset($taskForSave['picture_name']) && isset($taskForSave['picture_file_id'])) {
+            if (
+                isset($taskForSave['picture_name']) &&
+                isset($taskForSave['picture_file_id']) &&
+                $taskForSave['picture_name'] != $task->picture->name &&
+                $taskForSave['picture_file_id'] != $task->picture->file_id
+            ) {
                 $currentPicture = $task->picture;
                 if (isset($currentPicture)) {
+                    $googleDriveHelper = new GoogleDriveHelper(
+                        GoogleIdentityHelper::getGoogleClientWithToken(\Yii::$app->user->identity)
+                    );
+                    $googleDriveHelper->removeFile($currentPicture->file_id);
                     $currentPicture->delete();
                 }
